feature: update config
update babel to v7, mocha config files
