+++ frontend/modules/filter/types.js
@@ -34,6 +34,7 @@ const FILTER_KIND_LIST = [
     FILTER_KIND_TIME,
     FILTER_KIND_PRICE,
 ];
+
 export {
     SET_ONCHAIN_FILTER_PART,
     SET_REGULAR_FILTER_PART,
