+++ package.json
@@ -87,19 +87,33 @@
     }
   },
   "devDependencies": {
-    "babel": "^6.23.0",
-    "babel-core": "^6.25.0",
-    "babel-eslint": "^8.2.2",
-    "babel-loader": "^7.1.4",
+    "@babel/core": "^7.4.0",
+    "@babel/plugin-proposal-class-properties": "^7.4.0",
+    "@babel/plugin-proposal-decorators": "^7.4.0",
+    "@babel/plugin-proposal-do-expressions": "^7.2.0",
+    "@babel/plugin-proposal-export-default-from": "^7.2.0",
+    "@babel/plugin-proposal-export-namespace-from": "^7.2.0",
+    "@babel/plugin-proposal-function-bind": "^7.2.0",
+    "@babel/plugin-proposal-function-sent": "^7.2.0",
+    "@babel/plugin-proposal-json-strings": "^7.2.0",
+    "@babel/plugin-proposal-logical-assignment-operators": "^7.2.0",
+    "@babel/plugin-proposal-nullish-coalescing-operator": "^7.2.0",
+    "@babel/plugin-proposal-numeric-separator": "^7.2.0",
+    "@babel/plugin-proposal-optional-chaining": "^7.2.0",
+    "@babel/plugin-proposal-pipeline-operator": "^7.3.2",
+    "@babel/plugin-proposal-throw-expressions": "^7.2.0",
+    "@babel/plugin-syntax-dynamic-import": "^7.2.0",
+    "@babel/plugin-syntax-import-meta": "^7.2.0",
+    "@babel/plugin-transform-runtime": "^7.4.0",
+    "@babel/preset-env": "^7.4.2",
+    "@babel/preset-react": "^7.0.0",
+    "@babel/register": "^7.4.0",
+    "@babel/runtime": "^7.4.2",
+    "babel-eslint": "^9.0.0",
+    "babel-loader": "^8.0.5",
     "babel-plugin-istanbul": "^4.1.6",
     "babel-plugin-module-resolver": "^3.1.1",
-    "babel-plugin-transform-runtime": "^6.23.0",
-    "babel-polyfill": "^6.23.0",
-    "babel-preset-es2015": "^6.24.1",
-    "babel-preset-react": "^6.24.1",
-    "babel-preset-stage-0": "^6.24.1",
-    "babel-runtime": "^6.23.0",
-    "chai": "^3.5.0",
+    "chai": "^4.2.0",
     "cross-env": "^3.2.4",
     "electron": "^3.0.8",
     "electron-builder": "^20.28.3",
@@ -117,13 +131,14 @@
     "react-hot-loader": "^1.3.1",
     "redux-logger": "^2.10.2",
     "redux-mock-store": "^1.5.1",
-    "sinon": "^4.5.0",
-    "sinon-chai": "^3.0.0",
+    "sinon": "^7.3.0",
+    "sinon-chai": "^3.3.0",
     "spectron": "^3.8.0",
     "webpack": "4.12.1",
     "webpack-cli": "^3.0.8"
   },
   "dependencies": {
+    "@babel/polyfill": "^7.4.0",
     "@exponent/electron-cookies": "^2.0.0",
     "@grpc/proto-loader": "^0.3.0",
     "@journeyapps/sqlcipher": "^3.2.1",
@@ -140,6 +155,7 @@
     "moment": "^2.22.2",
     "moment-duration-format": "^2.2.2",
     "prop-types": "^15.6.0",
+    "raf": "^3.4.1",
     "rc-tooltip": "^3.4.2",
     "react": "^16.0.0",
     "react-addons-css-transition-group": "^15.6.2",
