+++ test/setup.js
@@ -1,19 +1,13 @@
 import chai from "chai";
 import sinon from "sinon";
 import sinonChai from "sinon-chai";
-
-const { JSDOM } = require("jsdom");
+import { JSDOM } from "jsdom";
 
 chai.use(sinonChai);
 const { expect } = chai;
 
 const jsdom = new JSDOM("<!doctype html><html><body></body></html>");
 const { window } = jsdom;
-
-const ipcClient = sinon.stub();
-
-// Functions
-window.ipcClient = ipcClient;
 
 // Constants
 window.env = process.env;
@@ -29,13 +23,43 @@ window.DB.Entities.Stream = require("../server/db/model/Stream").Stream;
 window.DB.Entities.StreamPart = require("../server/db/model/StreamPart").StreamPart;
 window.DB.Entities.Config = require("../server/db/model/Config").Config;
 
+// Functions
+const ipcClient = sinon.stub();
+window.ipcClient = ipcClient;
+
 const matchMedia = () => ({
     matches: false,
     addListener: () => {},
     removeListener: () => {},
 });
-
 window.matchMedia = window.matchMedia || matchMedia;
+
+// Base implementation of electron ipc. For test some code like window.ipcRenderer.on("lnd-down")
+let channels = {};
+const ipcRenderer = {
+    send: sinon.spy((channel, ...arg) => {
+        if (channel in channels) {
+            channels[channel].callbacks.forEach((callback) => {
+                if (arg.length) {
+                    callback(null, ...Object.values(arg[0]));
+                } else {
+                    callback(null);
+                }
+            });
+        }
+    }),
+    on: sinon.spy((channel, callback) => {
+        if (channel in channels) {
+            channels[channel].callbacks.push(callback);
+        } else {
+            channels[channel] = { callbacks: [callback] };
+        }
+    }),
+    reset: () => {
+        channels = {};
+    },
+};
+window.ipcRenderer = ipcRenderer;
 
 global.btoa = sinon.stub();
 global.window = window;
