+++ test/unit/modules/filter.spec.js
@@ -1,4 +1,5 @@
 import moment from "moment";
+
 import {
     filterActions as actions,
     filterTypes as types,
