+++ webpack.config.js
@@ -7,7 +7,7 @@ module.exports = {
     context: path.join(__dirname, "/frontend"),
 
     entry: {
-        bundle: "./app.js",
+        bundle: ["@babel/polyfill", "./app.js"],
     },
 
     output: {
@@ -34,11 +34,6 @@ module.exports = {
                 test: /\.jsx?$/,
                 exclude: /(node_modules|bower_components)/,
                 loader: "babel-loader",
-                options: {
-                    presets: ["es2015", "react", "stage-0"],
-                    plugins: ["transform-runtime"],
-                    retainLines: true,
-                },
             },
         ],
     },
